[refactor] 구독자 수 Long 타입 통일 및 구독 관련 조회 로직 ClubSubscriptionPort로 이동

+++ src/main/java/com/kustacks/kuring/club/adapter/in/web/dto/ClubDetailResponse.java
@@ -10,7 +10,7 @@ public record ClubDetailResponse(
         String summary,
         String category,
         String division,
-        int subscriberCount,
+        Long subscriberCount,
         boolean isSubscribed,
         String instagramUrl,
         String youtubeUrl,
